+++ src/trx/config/types.h
@@ -130,7 +130,12 @@ typedef struct {
         struct {
             BAR_LOCATION location;
             BAR_COLOR color;
-        } lara_health_bar, lara_air_bar, lara_sprint_bar, lara_exposure_bar;
+            BAR_COLOR poison_color;
+        } lara_health_bar;
+        struct {
+            BAR_LOCATION location;
+            BAR_COLOR color;
+        } lara_air_bar, lara_sprint_bar, lara_exposure_bar;
         struct {
             BAR_SHOW_MODE show_mode;
             BAR_LOCATION location;