+++ src/trx/game/lara/types.h
@@ -50,7 +50,7 @@ typedef struct {
     int16_t death_timer;
     int16_t sprint_timer;
     int16_t exposure_timer;
-    int16_t poison;
+    int16_t poison_timer;
     int32_t idle_timer;
     int16_t current_active;
     LOT_INFO lot;